+++ gammapy/background/__init__.py
@@ -1,5 +1,4 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Background estimation and modeling methods."""
-from .background_estimate import *
 from .phase import *
 from .reflected import *
