+++ gammapy/spectrum/__init__.py
@@ -1,5 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """1D spectrum analysis."""
+from .background_estimate import *
 from .core import *
 from .cosmic_ray import *
 from .crab import *
