+++ gammapy/spectrum/background_estimate.py
